# contracts/openapi.v1.json
{
  "components": {
    "schemas": {
      "ActivityClass": {
        "enum": [
          "process_spawn",
          "network_connect",
          "file_write"
        ],
        "type": "string"
      },
...
      "DivergenceFinding": {
        "properties": {
          "activity_class": {
            "$ref": "#/components/schemas/ActivityClass"
          },
          "detected_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DivergenceKind"
          },
          "observation_external_id": {
            "type": "string"
          },
          "receipt_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource": {
            "type": "string"
          },
          "sandbox_id": {
            "$ref": "#/components/schemas/SandboxId"
          },
          "session_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DivergenceFindingStatus"
          }
        },
        "required": [
          "id",
          "sandbox_id",
          "observation_external_id",
          "session_id",
          "kind",
          "activity_class",
          "resource",
          "status",
          "detected_at"
        ],
        "type": "object"
      },
...
      "DivergenceFindingStatus": {
        "enum": [
          "open",
          "resolved"
        ],
        "type": "string"
      },
      "DivergenceKind": {
        "enum": [
          "unaccounted_activity",
          "receipt_scope_mismatch"
        ],
        "type": "string"
      },
      "DivergenceReconcileRequest": {
        "properties": {
          "observations": {
            "items": {
              "$ref": "#/components/schemas/SensorObservation"
            },
            "type": "array"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "source",
          "observations"
        ],
        "type": "object"
      },
      "DivergenceReconcileResponse": {
        "properties": {
          "findings_created": {
            "items": {
              "$ref": "#/components/schemas/DivergenceFinding"
            },
            "type": "array"
          },
          "observations_ingested": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "observations_matched": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "ok": {
            "type": "boolean"
          },
          "retry_after": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "ok",
          "observations_ingested",
          "observations_matched",
          "findings_created"
        ],
        "type": "object"
      },
...
      "ReceiptScope": {
        "properties": {
          "activity_class": {
            "$ref": "#/components/schemas/ActivityClass"
          },
          "resource_prefix": {
            "type": "string"
          }
        },
        "required": [
          "activity_class",
          "resource_prefix"
        ],
        "type": "object"
      },
...
      "SandboxId": {
        "format": "uuid",
        "type": "string"
      },
...
      "SensorObservation": {
        "properties": {
          "activity_class": {
            "$ref": "#/components/schemas/ActivityClass"
          },
          "external_id": {
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "sandbox_id": {
            "$ref": "#/components/schemas/SandboxId"
          },
          "session_id": {
            "type": "string"
          }
        },
        "required": [
          "external_id",
          "sandbox_id",
          "session_id",
          "activity_class",
          "resource",
          "observed_at"
        ],
        "type": "object"
      },
...
      "ToolCallLedgerEntryRequest": {
        "properties": {
          "ended_at": {
            "format": "date-time",
            "type": "string"
          },
          "external_id": {
            "type": "string"
          },
          "receipt_id": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/ReceiptScope"
            },
            "type": "array"
          },
          "session_id": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "external_id",
          "session_id",
          "receipt_id",
          "started_at",
          "ended_at",
          "scopes"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "EvalOps contributors"
    },
    "description": "HTTP control plane for disposable development sandboxes",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Sandboxwich API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/divergence/reconcile": {
      "post": {
        "operationId": "reconcile_divergence",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DivergenceReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivergenceReconcileResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "divergence"
        ]
      }
    },
...
    "/v1/sandboxes/{sandbox_id}/tool-call-ledger": {
      "post": {
        "operationId": "append_tool_call_ledger",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCallLedgerEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "divergence"
        ]
      }
    },
...
  },
...
}